+++ backend/src/sglportalapi/tests.py
@@ -9,8 +9,8 @@ from datetime import date
 from pathlib import Path
 from typing import List, Dict, Optional
 
-
-from sglportalapi.clientside import check_session_archive
+from sglportalapi import PL2
+from sglportalapi.clientside import check_session_archive, get_trial_timing_from_pl2_file
 from sglportalapi.maestro import Protocol, DataFileHeader
 
 
@@ -97,6 +97,61 @@ def list_units_in_archive(zip_file_path: str) -> None:
         if not found:
             print(f"==> Did not find a pickle file with neural unit info in the archive.")
 
+def show_pl2_info_in_archive(zip_file_path: str) -> None:
+    """
+    If the specified archive contains one or more PL2 files, display relevant information recorded therein: WB<n> and
+    SPKC<n> channels that were recorded; Maestro trial timing information extracted from the PL2 'Strobed' and 'EVT02'
+    channels.
+
+    This method will take a significant amount of time to do its work. Since the PL2 files are typically multi-GB and
+    the information is scattered throughout, the most efficient way to glean the information is to extract the PL2 file
+    from the archive, then delete the extracted file afterwards.
+    """
+    try:
+        zip_path = Path(zip_file_path)
+        with zipfile.ZipFile(zip_path, 'r') as archive:
+            pl2_files: List[zipfile.ZipInfo] = list()
+            for archive_info in archive.infolist():
+                if (len(archive_info.filename) > 4) and (archive_info.filename[-4:].lower() == '.pl2'):
+                    pl2_files.append(archive_info)
+            if len(pl2_files) == 0:
+                raise Exception("Did not find a PL2 file in the archive.")
+
+            for pl2_zip_info in pl2_files:
+                pl2_temp: Optional[Path] = None
+                try:
+                    print("Extracting PL2 file... this will take a while.")
+                    path_str = archive.extract(pl2_zip_info, zip_path.parent)
+                    pl2_temp = Path(path_str)
+                    print(f"Done extracting. File at {path_str}")
+
+                    with open(pl2_temp, 'rb') as fp:
+                        print(f"Processing trial timing information in {pl2_zip_info.filename}...")
+                        info = PL2.load_file_information(fp)
+                        timings_dict = get_trial_timing_from_pl2_file(fp, info)
+                        for k, v in timings_dict.items():
+                            print(f"   {k}: start={v[0]:.3f}, stop={v[1]:.3f}")
+
+                        print(f"\nRecorded wideband or narrowband analog channels in {pl2_zip_info.filename}:")
+                        channel_list = info['analog_channels']
+                        for i in range(len(channel_list)):
+                            if channel_list[i]['num_values'] > 0:
+                                if channel_list[i]['source'] in [PL2.PL2_ANALOG_TYPE_WB, PL2.PL2_ANALOG_TYPE_SPKC]:
+                                    num_blocks = len(channel_list[i]["block_num_items"])
+                                    samples_per_sec: float = channel_list[i]['samples_per_second']
+                                    n_samples = channel_list[i]['num_values']
+                                    src = "WB" if channel_list[i]['source'] == PL2.PL2_ANALOG_TYPE_WB else "SPKC"
+                                    ch_label = f"{src}{str(channel_list[i]['channel'])}"
+                                    print(f"   {ch_label}: N={n_samples}, rate={samples_per_sec:.1f} Hz, "
+                                          f"blocks={num_blocks}")
+                finally:
+                    if isinstance(pl2_temp, Path) and pl2_temp.is_file():
+                        pl2_temp.unlink(missing_ok=True)
+                        print("Extracted PL2 file was removed")
+    except Exception as e:
+        print(f"ERROR: {str(e)}")
+
+
 
 def _print_usage() -> None:
     print("\nAvailable commands:\n"
@@ -105,6 +160,7 @@ def _print_usage() -> None:
           "   c <path> = Perform sanity check on contents of the session archive at <path>.\n"
           "   t <path> = List all unique trial protocols found in the session archive at <path>.\n"
           "   u <path> = List channel and spike train length for each neural unit found in archive at <path>.\n"
+          "   p <path> = Show analog channel and Maestro trial timing info from PL2 file in archive at <path>.\n"
           "   h = Print this usage message.\n"
           "   q = Quit.\n\n", file=sys.stdout, flush=True)
 
@@ -119,13 +175,14 @@ def _process_command() -> bool:
         _print_usage()
     elif parts[0] == 'q':
         quit_requested = True
-    elif (parts[0] in ['l', 'c', 'd', 't', 'u']) and (len(parts) == 2):
+    elif (parts[0] in ['l', 'c', 'd', 't', 'u', 'p']) and (len(parts) == 2):
         if not Path(parts[1]).is_file():
             print(f"Archive file not found: {parts[1]}\n\n")
         elif parts[0] == 'l':
             list_files_in_archive(parts[1])
             print("\n\n")
         elif parts[0] == 'c':
+            print(f"PLEASE WAIT -- This sanity check will take a while!")
             err_msg = check_session_archive(parts[1])
             if len(err_msg) > 0:
                 print(f"{err_msg}\n\n")
@@ -137,8 +194,11 @@ def _process_command() -> bool:
         elif parts[0] == 't':
             list_protocols(parts[1])
             print("\n\n")
-        else:
+        elif parts[0] == 'u':
             list_units_in_archive(parts[1])
+            print("\n\n")
+        else:
+            show_pl2_info_in_archive(parts[1])
             print("\n\n")
     else:
         print(f"Invalid command. Try again.\n\n")